Do not block Submit() when the HMD is removed.
When the OpenXR session is not active, do not block and wait for it to become
active - instead silently discard the application's frames. This is
required for some runtimes (notably on the Oculus Quest) that won't
deliver a state-change event until some function in the main loop is
called.

It also mirrors the behaviour of SteamVR - when the HMD is removed the
game keeps rendering unless it decides not to based on the HMD status.