Add the soft abort system
For the last almost-five-years, aborts (particularly in the form of the dreaded
STUBBED() call) have been a major pain in the backside of most OpenComposite
users.

The primary reason that there are so many assertions in OpenComposite that
probably could have let the game try and continue degraded is that there's
always going to be a few games when doing that will cause a crash that's only
found after several days of reverse-engineering, whereas just fixing the aborts
as they're thrown is much quicker since I know where to look.

Unfortunately this is obviously painful to everyone trying to use it, since it
makes it dramatically more likely that users will experience a crash that they
didn't need to.

The solution here (and honestly I really should have done this way earlier) is
to have introduce a new abort call that normally logs a single (irrespective of
how many times they're called to avoid logspam) stern warning, but continues
anyway and allows OC to return an error code. However, an option in the config
causes these functions to call the regular abort instead, exiting the programme.

That way, you can get a lead on what's likely causing the crash without
excessive debugging, while increasing end-user reliability.

It's probably not appropriate to s/OOVR_ABORTF/OOVR_SOFT_ABORTF/ since there's
a lot of places where a crash is certain (for example, an assertion checking a
pointer is non-null before dereferencing it) and in those cases showing a popup
clearly attributing it to OC is better than letting the game crash silently
with no indication of the cause, least the devs get blamed for it - but over
time these cases can be chiseled away at, using the openvr-tests project to
perfectly replicate SteamVR's behaviour with abnormal arguments.