Better handling of timing issues on xrLocate functions.
xrLocate functions may return XR_ERROR_TIME_INVALID if the time requested is outside some arbitrary value set by the runtime:
https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#prediction-time-limits
Usually this is just due to lagging on loading screens so should be able to continue with pose set to invalid. Some runtimes (Oculus!) seem to return other failure codes for this situation but can still carry on.