Make InputValueHandle behave like XrPaths
Previously a vr::InputValueHandle_t (henceforth IVH) would represent a given
device index. This is absolutely not how SteamVR works, and will cause lots
of issues in applications that depend on this working properly.

This makes the handles pointers to small structs, and allows the user to
construct them for arbitrary paths, regardless of if we associate a meaning
with them or not.