Remove "using namespace std"
This is fairly widely considered bad practice, and when using MSVC with
/std:c++20 it causes the build to fail, due to "byte" being defined in
the C++ standard library and a Windows header.

There is a define to fix that, but removing the global std import is
good practice anyway.

This patch still imports string and to_string (and their wide versions)
since they're super frequently used and it is convenient not to type
std:: in front of them, along with saving a lot of time when making this
patch.