Fix BindInputsForSession crashing if inputs haven't been loaded
Actions and action sets are attached to the XrInstance, however a few
things (action spaces, hand tracking, and selecting the action sets to
enable) are attached to sessions. Since we have to restart the session
when we find out what graphics API the application is using, this
per-session binding logic is separated into the BindInputsForSession
method.

This method is called both when inputs are first loaded, and by
DrvOpenXR when it restarts the session. Previously if the BaseInput
class had been created and the session restarted, this method was always
called. This was generally fine with both legacy input-based games
(where BaseInput presumably hadn't been created) and with SteamVR Input
based games (which would initialise BaseInput early on) but can fail if
they call some methods on BaseInput but don't load the actions it before
submitting their first frame.

When this happened, BindInputsForSession would attempt to create action
spaces for XR_NULL actions, leading to an assertion failure.

This patch checks if the actions have been loaded, and if not doesn't do
anything.