Soft abort for Prop_UserHeadToEyeDepthMeters_Float
This property is for eye-relief and as per Ludwig in the linked OpenVR
issue, "[is] useful is certain unusual circumstances". Setting it to
zero is the defualt.

This puts in a soft abort since we used to get it from hmdToEyeViewPose
in LibOVR, but I'm not sure whether that was ever correct.