Disable GLM rgba/stpq accessors
Previously, you could access a GLM vector as v.r or v.s instead of v.x -
this was done using a union, which unfortunately made vectors,
quaternions and matrices all a big pain to look at in the CLion debugger
as you had to manually expand the union for each value. Matrices were a
particularly big pain.

I suspect this also helps in the Visual Studio debugger, but haven't
checked.