Rewrite the (model-space) hand tracking processing
Previously the hand-tracking system apparently worked in Neos and a
couple of other games, but when rendering with the test application the
bones were broken. This is likely because the test application shows the
raw pose, while most games place limits on the hands to avoid them doing
anythign too stupid.

This rewrite is significantly shorter, hopefully easier to understand,
and gives a result very close (identical?) to SteamVR in the test
application, when the latter is modified to always ask for model-space
poses.