Use the correct-sized swapchain for OpenGL
Previously, we'd make the OpenGL swapchain the same size as the source
image. This was extremely inefficient for games that render their eyes
side-by-side to a single texture, as we'd have two full-sized
swapchains.

This patch means we'll make a swapchain matching the size of the image
the game is submitting, copy the portion of the image the game selected,
and mark the entire swapchain as being visible.