OpenGL fix for gamma-corrected RGBA8 format
Applications that store color data from the gamma-corrected color space in a GL_RGBA8 linear format will see incorrect color output. The displayed imaged will have incorrect colors and the gamma will be too high.

While this is an unusual situation, it is allowed by the standard. The GLCompositor has been modified to check if a received texture is using a linear RGBA8 format with the EColorSpace set to ColorSpace_Gamma. In this case, the format is swapped out for SRGB8_ALPHA8. This tells the downstream render pipeline that the texture data has gamma-corrected color, so it can be handled correctly.