Fix Oculus v59 compatibility
Wait for the XR_SESSION_STATE_EXITING session state before destroying the session.  This is the session state that it reaches and stays at when ending the session.  It doesn't seem to return to Idle.

The problem that started occurring in Oculus v59 was that sessions had to be fully gracefully ended with xrRequestExitSession and xrEndSession before the Oculus runtime would release the system id to be used again.  Any attempt to use that system id while not in the original session would fail, saying it was an invalid system id.  Outright destroying the session did not work, as the Oculus runtime would not release the system id in that case.  Only fully ending the session before destroying it triggers the system id release.

The session state reaches 8, which is XR_SESSION_STATE_EXITING when the session is ended and it stays there.  Once it reaches this state, it seems to be safe to destroy the session.  So a loop was added in the session shut down to verify that it has reached XR_SESSION_STATE_EXITING before destroying the session and recreating the system id.